Tweak Manned Contracts
Update the manned contracts to require new vessels and the same vessel to return home.
Fix the manned landing contracts where the landing param was disableOnStateChange = false..
Fix the manned long term orbit contract so the return home param was not part of the 7d duration vessel param group
Change the EVA contract to require just an eva and not an eva report.